requirements.txt
flask
flask-babel
email_validator
Pillow
python-dotenv
gunicorn
authlib
requests
